LPS-96279 ContributorBuilder won't fetch for members
Fetching to get the number of members is perform one level up by the
SegmentEdit component. ContributorBuilder consumes the state related to
that operation as props. This makes ContributorBuilder component more
flexible.
